ai/pyproject.toml
[tool.poetry]
name = "hw"
version = "0.1.0"
description = ""
authors = ["Tanmai Kalisipudi <tanmai.kalisipudi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
flask-cors = "^5.0.0"
supabase = "^2.10.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.12"
langchain-community = "^0.3.11"
langgraph = "^0.2.58"
langchain-anthropic = "^0.3.0"
tavily-python = "^0.5.0"
langgraph-checkpoint-sqlite = "^2.0.1"
browser-use = "^0.1.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
